First pass at old client and server
The server will listen on localhost on a provided port (default 8080) and `/envelope`
route for v1 envelopes, unmarshal and remarshal them, then sending them
back.

The client will send a v1 envelope to localhost, the provided port
(default 8080) and `/envelope` route. It will attempt to unmarshal the
response as an envelope.